+++ berkeley-function-call-leaderboard/bfcl/model_handler/handler_map.py
@@ -7,26 +7,25 @@ from bfcl.model_handler.oss_model.hammer import HammerHandler
 from bfcl.model_handler.oss_model.hermes import HermesHandler
 from bfcl.model_handler.oss_model.llama import LlamaHandler
 from bfcl.model_handler.oss_model.llama_fc import LlamaFCHandler
-from bfcl.model_handler.oss_model.phi import PhiHandler
-from bfcl.model_handler.oss_model.salesforce import SalesforceHandler
-from bfcl.model_handler.oss_model.qwen import QwenHandler
 from bfcl.model_handler.oss_model.minicpm import MiniCPMHandler
+from bfcl.model_handler.oss_model.phi import PhiHandler
+from bfcl.model_handler.oss_model.qwen import QwenHandler
+from bfcl.model_handler.oss_model.salesforce import SalesforceHandler
 from bfcl.model_handler.proprietary_model.claude import ClaudeHandler
 from bfcl.model_handler.proprietary_model.cohere import CohereHandler
 from bfcl.model_handler.proprietary_model.databricks import DatabricksHandler
 from bfcl.model_handler.proprietary_model.fireworks import FireworksHandler
 from bfcl.model_handler.proprietary_model.functionary import FunctionaryHandler
 from bfcl.model_handler.proprietary_model.gemini import GeminiHandler
+from bfcl.model_handler.proprietary_model.gogoagent import GoGoAgentHandler
 from bfcl.model_handler.proprietary_model.gorilla import GorillaHandler
 from bfcl.model_handler.proprietary_model.mistral import MistralHandler
 from bfcl.model_handler.proprietary_model.nexus import NexusHandler
 from bfcl.model_handler.proprietary_model.nvidia import NvidiaHandler
 from bfcl.model_handler.proprietary_model.openai import OpenAIHandler
 from bfcl.model_handler.proprietary_model.yi import YiHandler
-from bfcl.model_handler.proprietary_model.deepseekv2 import DeepseekV2Handler
-from bfcl.model_handler.proprietary_model.gogoagent import GoGoAgentHandler
 
-# TODO: Add Deepseek V2, meta-llama/Llama-3.1-405B-Instruct
+# TODO: Add meta-llama/Llama-3.1-405B-Instruct
 
 # Inference through API calls
 api_inference_handler_map = {
@@ -129,12 +128,10 @@ local_inference_handler_map = {
     "Qwen/Qwen2.5-7B-Instruct": QwenHandler,
     "Team-ACE/ToolACE-8B": LlamaHandler,
     "openbmb/MiniCPM3-4B": MiniCPMHandler,
-    "deepseek-ai/DeepSeek-V2-Chat-0628": DeepseekV2Handler,
-    "deepseek-ai/DeepSeek-V2-Chat": DeepseekV2Handler,
-    "deepseek-ai/DeepSeek-V2": DeepseekV2Handler,
-    "deepseek-ai/DeepSeek-V2-Lite": DeepseekV2Handler,
-    "deepseek-ai/DeepSeek-V2-Lite-Chat": DeepseekV2Handler,
-    "deepseek-ai/-Coder-V2-Instruct": DeepseekHandler,
+    "deepseek-ai/DeepSeek-Coder-V2-Instruct": DeepseekHandler,
+    "deepseek-ai/DeepSeek-Coder-V2-Lite-Instruct": DeepseekHandler,
+    "deepseek-ai/DeepSeek-V2": DeepseekHandler,
+    "deepseek-ai/DeepSeek-V2-Lite": DeepseekHandler,
 }
 
 # Deprecated/outdated models, no longer on the leaderboard